+++ lib/restrr.dart
@@ -1,9 +1,25 @@
 library;
 
-export 'src/restrr_base.dart';
+/* [ /src ] */
 export 'src/entity_builder.dart';
-export 'src/entities/status/health_response.dart';
-export 'src/requests/error_response.dart';
-export 'src/requests/rest_response.dart';
-export 'src/requests/routes.dart';
-export 'src/util/string_utils.dart';
+export 'src/restrr_base.dart';
+
+/* [ /src/entities ] */
+export 'src/entities/health_response.dart';
+export 'src/entities/restrr_entity.dart';
+export 'src/entities/user.dart';
+
+/* [ /src/requests ] */
+export 'src/requests/route.dart';
+export 'src/requests/route_definitions.dart';
+
+/* [ /src/requests/responses ] */
+export 'src/requests/responses/error_response.dart';
+export 'src/requests/responses/rest_response.dart';
+
+/* [ /src/service ] */
+export 'src/service/api_service.dart';
+export 'src/service/user_service.dart';
+
+/* [ /src/utils ] */
+export 'src/utils/string_utils.dart';