+++ lib/src/entity_builder.dart
@@ -1,5 +1,6 @@
 import 'package:restrr/restrr.dart';
 
+/// Defines how to build entities from JSON responses.
 class EntityBuilder {
   final RestrrImpl api;
 
@@ -10,6 +11,17 @@ class EntityBuilder {
       healthy: json['healthy'],
       apiVersion: json['api_version'],
       details: json.keys.contains('details') ? json['details'] : null,
+    );
+  }
+
+  User buildUser(Map<String, dynamic> json) {
+    return UserImpl(
+      api: api,
+      id: json['id'],
+      username: json['username'],
+      email: json['email'],
+      createdAt: DateTime.parse(json['created_at']),
+      isAdmin: json['is_admin'],
     );
   }
 }