+++ lib/src/requests/responses/rest_response.dart
@@ -1,5 +1,7 @@
-import '../../restrr.dart';
+import '../../../restrr.dart';
 
+/// Represents a response from a REST request.
+/// This can either hold data, [T], or an [ErrorResponse].
 class RestResponse<T> {
   final T? data;
   final ErrorResponse? error;