+++ lib/src/utils/string_utils.dart
@@ -1,6 +1,7 @@
 class StringUtils {
   const StringUtils._();
 
+  /// Returns the number of occurrences of [search] in [str].
   static int count(String str, String search) {
     return search.allMatches(str).length;
   }